extract cache & digest
